requirements.txt
django
djangorestframework
gunicorn # WAS server
